agent_deploy/requirements.txt
fastapi
uvicorn==0.24.0
firebase-admin==6.2.0
redis
rq
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.4
browser-use
langchain==0.3.14
langchain-openai==0.3.1
playwright==1.49.0